refactor: add configurable multi protocol provider
