+++ typescript/sdk/src/providers/MultiProtocolProvider.ts
@@ -1,302 +1,77 @@
-import { Logger } from 'pino';
-
-import {
-  Address,
-  HexString,
-  ProtocolType,
-  objFilter,
-  objMap,
-  pick,
-  rootLogger,
-} from '@hyperlane-xyz/utils';
+import { ProtocolType } from '@hyperlane-xyz/utils';
 
 import { multiProtocolTestChainMetadata } from '../consts/testChains.js';
-import { ChainMetadataManager } from '../metadata/ChainMetadataManager.js';
 import type { ChainMetadata } from '../metadata/chainMetadataTypes.js';
-import type { ChainMap, ChainName, ChainNameOrId } from '../types.js';
+import type { ChainMap, ChainName } from '../types.js';
 
-import { MultiProvider, MultiProviderOptions } from './MultiProvider.js';
 import {
-  AleoProvider,
-  CosmJsNativeProvider,
-  CosmJsProvider,
-  CosmJsWasmProvider,
-  EthersV5Provider,
-  PROTOCOL_TO_DEFAULT_PROVIDER_TYPE,
-  ProviderMap,
-  ProviderType,
-  RadixProvider,
-  SolanaWeb3Provider,
-  StarknetJsProvider,
-  TronProvider,
-  TypedProvider,
-  TypedTransaction,
-  ViemProvider,
-} from './ProviderType.js';
+  ConfiguredMultiProtocolProvider,
+  ConfiguredMultiProtocolProviderOptions,
+} from './ConfiguredMultiProtocolProvider.js';
+import { MultiProvider } from './MultiProvider.js';
+import { ProviderType, TypedProvider } from './ProviderType.js';
 import {
+  ProviderBuilderFn,
   ProviderBuilderMap,
   defaultProviderBuilderMap,
   defaultTronEthersProviderBuilder,
 } from './providerBuilders.js';
-import {
-  TransactionFeeEstimate,
-  estimateTransactionFee,
-} from './transactionFeeEstimators.js';
 
-export interface MultiProtocolProviderOptions {
-  logger?: Logger;
-  providers?: ChainMap<ProviderMap<TypedProvider>>;
-  providerBuilders?: Partial<ProviderBuilderMap>;
-}
+export interface MultiProtocolProviderOptions extends ConfiguredMultiProtocolProviderOptions {}
 
-/**
- * A version of MultiProvider that can support different
- * provider types across different protocol types.
- *
- * This uses a different interface for provider/signer related methods
- * so it isn't strictly backwards compatible with MultiProvider.
- *
- * Unlike MultiProvider, this class does not support signer/signing methods (yet).
- * @typeParam MetaExt - Extra metadata fields for chains (such as contract addresses)
- */
 export class MultiProtocolProvider<
   MetaExt = {},
-> extends ChainMetadataManager<MetaExt> {
-  // Chain name -> provider type -> provider
-  protected readonly providers: ChainMap<ProviderMap<TypedProvider>>;
-  // Chain name -> provider type -> signer
-  protected signers: ChainMap<ProviderMap<never>> = {}; // TODO signer support
-  protected readonly providerBuilders: Partial<ProviderBuilderMap>;
-  public readonly logger: Logger;
-
-  constructor(
-    chainMetadata: ChainMap<ChainMetadata<MetaExt>>,
-    protected readonly options: MultiProtocolProviderOptions = {},
-  ) {
-    super(chainMetadata, options);
-    this.logger =
-      options?.logger ||
-      rootLogger.child({
-        module: 'MultiProtocolProvider',
-      });
-    this.providers = options.providers || {};
-    this.providerBuilders =
-      options.providerBuilders || defaultProviderBuilderMap;
-  }
-
+> extends ConfiguredMultiProtocolProvider<MetaExt> {
   static fromMultiProvider<MetaExt = {}>(
     mp: MultiProvider<MetaExt>,
     options: MultiProtocolProviderOptions = {},
   ): MultiProtocolProvider<MetaExt> {
     const newMp = new MultiProtocolProvider<MetaExt>(mp.metadata, options);
-
-    const typedProviders = objMap(mp.providers, (_, provider) => ({
-      type: ProviderType.EthersV5,
-      provider,
-    })) as ChainMap<TypedProvider>;
-
-    newMp.setProviders(typedProviders);
+    newMp.setProviders(
+      Object.fromEntries(
+        Object.entries(mp.providers).map(([chain, provider]) => [
+          chain,
+          {
+            type: ProviderType.EthersV5,
+            provider,
+          },
+        ]),
+      ) as ChainMap<TypedProvider>,
+    );
     return newMp;
   }
 
-  toMultiProvider(options?: MultiProviderOptions): MultiProvider<MetaExt> {
-    const newMp = new MultiProvider<MetaExt>(this.metadata, options);
+  constructor(
+    chainMetadata: ChainMap<ChainMetadata<MetaExt>>,
+    options: MultiProtocolProviderOptions = {},
+  ) {
+    super(chainMetadata, {
+      ...options,
+      providerBuilders: options.providerBuilders || defaultProviderBuilderMap,
+    });
+  }
 
-    const providers = objMap(
-      this.providers,
-      (_, typeToProviders) => typeToProviders[ProviderType.EthersV5]?.provider,
-    ) as ChainMap<EthersV5Provider['provider'] | undefined>;
-
-    const filteredProviders = objFilter(
-      providers,
-      (_, p): p is EthersV5Provider['provider'] => !!p,
-    ) as ChainMap<EthersV5Provider['provider']>;
-
-    newMp.setProviders(filteredProviders);
-    return newMp;
+  protected override getProviderBuilder(
+    protocol: ProtocolType,
+    type: ProviderType,
+  ): ProviderBuilderFn<TypedProvider> | undefined {
+    if (protocol === ProtocolType.Tron && type === ProviderType.EthersV5) {
+      return (urls, network) => ({
+        type: ProviderType.EthersV5,
+        provider: defaultTronEthersProviderBuilder(urls, network),
+      });
+    }
+    return this.providerBuilders[type];
   }
 
   override extendChainMetadata<NewExt = {}>(
     additionalMetadata: ChainMap<NewExt>,
   ): MultiProtocolProvider<MetaExt & NewExt> {
     const newMetadata = super.extendChainMetadata(additionalMetadata).metadata;
-    const newMp = new MultiProtocolProvider(newMetadata, {
+    return new MultiProtocolProvider(newMetadata, {
       ...this.options,
       providers: this.providers,
-    });
-    return newMp;
-  }
-
-  tryGetProvider(
-    chainNameOrId: ChainNameOrId,
-    type?: ProviderType,
-  ): TypedProvider | null {
-    const metadata = this.tryGetChainMetadata(chainNameOrId);
-    if (!metadata) return null;
-    const { protocol, name, chainId, rpcUrls } = metadata;
-    // Cannot build provider for unknown protocol types (forward compatibility)
-    if (protocol === ProtocolType.Unknown) return null;
-    type = type || PROTOCOL_TO_DEFAULT_PROVIDER_TYPE[protocol];
-    if (!type) return null;
-
-    if (this.providers[name]?.[type]) return this.providers[name][type]!;
-
-    // Tron chains requesting EthersV5 must use TronJsonRpcProvider instead of
-    // HyperlaneSmartProvider, which doesn't handle Tron's JSON-RPC quirks.
-    let builder;
-    if (protocol === ProtocolType.Tron && type === ProviderType.EthersV5) {
-      builder = (urls: ChainMetadata['rpcUrls'], network: number | string) => ({
-        type: ProviderType.EthersV5 as const,
-        provider: defaultTronEthersProviderBuilder(urls, network),
-      });
-    } else {
-      builder = this.providerBuilders[type];
-    }
-    if (!rpcUrls.length || !builder) return null;
-
-    const provider = builder(rpcUrls, chainId);
-    this.providers[name] ||= {};
-    this.providers[name][type] = provider;
-    return provider;
-  }
-
-  getProvider(
-    chainNameOrId: ChainNameOrId,
-    type?: ProviderType,
-  ): TypedProvider {
-    const provider = this.tryGetProvider(chainNameOrId, type);
-    if (!provider)
-      throw new Error(`No provider available for ${chainNameOrId}`);
-    return provider;
-  }
-
-  protected getSpecificProvider<T>(
-    chainNameOrId: ChainNameOrId,
-    type: ProviderType,
-  ): T {
-    const provider = this.getProvider(chainNameOrId, type);
-    if (provider.type !== type)
-      throw new Error(
-        `Invalid provider type, expected ${type} but found ${provider.type}`,
-      );
-    return provider.provider as T;
-  }
-
-  getEthersV5Provider(
-    chainNameOrId: ChainNameOrId,
-  ): EthersV5Provider['provider'] {
-    return this.getSpecificProvider<EthersV5Provider['provider']>(
-      chainNameOrId,
-      ProviderType.EthersV5,
-    );
-  }
-
-  getViemProvider(chainNameOrId: ChainNameOrId): ViemProvider['provider'] {
-    return this.getSpecificProvider<ViemProvider['provider']>(
-      chainNameOrId,
-      ProviderType.Viem,
-    );
-  }
-
-  getSolanaWeb3Provider(
-    chainNameOrId: ChainNameOrId,
-  ): SolanaWeb3Provider['provider'] {
-    return this.getSpecificProvider<SolanaWeb3Provider['provider']>(
-      chainNameOrId,
-      ProviderType.SolanaWeb3,
-    );
-  }
-
-  getCosmJsProvider(chainNameOrId: ChainNameOrId): CosmJsProvider['provider'] {
-    return this.getSpecificProvider<CosmJsProvider['provider']>(
-      chainNameOrId,
-      ProviderType.CosmJs,
-    );
-  }
-
-  getCosmJsWasmProvider(
-    chainNameOrId: ChainNameOrId,
-  ): CosmJsWasmProvider['provider'] {
-    return this.getSpecificProvider<CosmJsWasmProvider['provider']>(
-      chainNameOrId,
-      ProviderType.CosmJsWasm,
-    );
-  }
-
-  getCosmJsNativeProvider(
-    chainNameOrId: ChainNameOrId,
-  ): CosmJsNativeProvider['provider'] {
-    return this.getSpecificProvider<CosmJsNativeProvider['provider']>(
-      chainNameOrId,
-      ProviderType.CosmJsNative,
-    );
-  }
-
-  getStarknetProvider(
-    chainNameOrId: ChainNameOrId,
-  ): StarknetJsProvider['provider'] {
-    return this.getSpecificProvider<StarknetJsProvider['provider']>(
-      chainNameOrId,
-      ProviderType.Starknet,
-    );
-  }
-
-  getRadixProvider(chainNameOrId: ChainNameOrId): RadixProvider['provider'] {
-    return this.getSpecificProvider<RadixProvider['provider']>(
-      chainNameOrId,
-      ProviderType.Radix,
-    );
-  }
-
-  getAleoProvider(chainNameOrId: ChainNameOrId): AleoProvider['provider'] {
-    return this.getSpecificProvider<AleoProvider['provider']>(
-      chainNameOrId,
-      ProviderType.Aleo,
-    );
-  }
-
-  getTronProvider(chainNameOrId: ChainNameOrId): TronProvider['provider'] {
-    return this.getSpecificProvider<TronProvider['provider']>(
-      chainNameOrId,
-      ProviderType.Tron,
-    );
-  }
-
-  setProvider(
-    chainNameOrId: ChainNameOrId,
-    provider: TypedProvider,
-  ): TypedProvider {
-    const chainName = this.getChainName(chainNameOrId);
-    this.providers[chainName] ||= {};
-    this.providers[chainName][provider.type] = provider;
-    return provider;
-  }
-
-  setProviders(providers: ChainMap<TypedProvider>): void {
-    for (const chain of Object.keys(providers)) {
-      this.setProvider(chain, providers[chain]);
-    }
-  }
-
-  estimateTransactionFee({
-    chainNameOrId,
-    transaction,
-    sender,
-    senderPubKey,
-  }: {
-    chainNameOrId: ChainNameOrId;
-    transaction: TypedTransaction;
-    sender: Address;
-    senderPubKey?: HexString;
-  }): Promise<TransactionFeeEstimate> {
-    const provider = this.getProvider(chainNameOrId, transaction.type);
-    const chainMetadata = this.getChainMetadata(chainNameOrId);
-    return estimateTransactionFee({
-      transaction,
-      provider,
-      chainMetadata,
-      sender,
-      senderPubKey,
+      providerBuilders: this.providerBuilders as Partial<ProviderBuilderMap>,
     });
   }
 
@@ -308,16 +83,20 @@ export class MultiProtocolProvider<
     result: MultiProtocolProvider<MetaExt>;
   } {
     const { intersection, result } = super.intersect(chains, throwIfNotSubset);
-    const multiProvider = new MultiProtocolProvider(result.metadata, {
-      ...this.options,
-      providers: pick(this.providers, intersection),
-    });
-    return { intersection, result: multiProvider };
+    return {
+      intersection,
+      result: new MultiProtocolProvider(result.metadata, {
+        ...this.options,
+        providers: Object.fromEntries(
+          Object.entries(this.providers).filter(([chain]) =>
+            intersection.includes(chain),
+          ),
+        ),
+        providerBuilders: this.providerBuilders,
+      }),
+    };
   }
 
-  /**
-   * Creates a MultiProvider for test networks
-   */
   static createTestMultiProtocolProvider<MetaExt = {}>(
     metadata = multiProtocolTestChainMetadata,
     providers: Partial<Record<ProtocolType, TypedProvider>> = {},
